Structured light wordpress theme
Structured is a light wordpress theme
based on bootstrap 5.

It is built with the idea of implementing basic functions in functions.php so to  avoid the need of many small plugins

Its layout is based on bootstrap 5.
It has Font Awesome 5 inside

See in functions.php (lines 3-12) the  functions included and comment the ones you are not using.
-> svg support
-> custom-login page
-> disable wordpress api
-> disable comments
-> duplicate post feature
-> social share shortcode: use [Social_Share]

See from wordpress admin Appearance->Theme Options to activate:
-> google analytics/matomo/metricool /fb Pixel
-> Default OG:image used where no specified in "featured post image"